requirements.txt
Django==1.7.5
Fabric==1.10.1
Jinja2==2.7.3
Markdown==2.5.2
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.7.0
Pygments==2.0.1
South==1.0.1
Unidecode==0.04.16
blinker==1.3
django-ace==1.0.2
django-autocomplete-light==2.0.3
django-autoslug==1.7.2
django-cities-light==3.0.2
django-filebrowser==3.5.7
django-grappelli==2.6.3
django-jinja==1.0.5
django-taggit==0.12.3
djangorestframework==3.0.1
docutils==0.12
ecdsa==0.11
feedgenerator==1.7
markdown2==2.3.0
paramiko==1.15.2
pelican==3.5.0
pycrypto==2.6.1
python-dateutil==2.3
pytils==0.3
pytz==2014.10
six==1.8.0
sorl-thumbnail==12.2
trans==2.0.1
ujson==1.33
wsgiref==0.1.2
